requirements.txt
selenium
webdriver-manager
pyyaml
